refactor(55e): 搬移根目錄 test_frontend_lint.py → tests/unit/，刪除根檔
依 plan §T55e + TASK-55e.md：
- 將 tests/test_frontend_lint.py（48 classes / 163 tests）全部搬移到
  tests/unit/test_frontend_lint.py 末尾（# === 從 tests/test_frontend_lint.py
  搬移（T55e）=== 分隔註解之後）
- 刪除根目錄檔案，達成 plan §3 整合驗收「root 檔不再存在」

unit 檔頂部新增（從 root 帶入，去重）：
- from typing import List, Tuple
- PROJECT_ROOT = Path(__file__).parent.parent.parent
- VANILLA_HANDLER_PATTERN regex
- find_pattern_in_file() helper

搬移過程修正 2 個潛在 bug（從 tests/ 跨 tests/unit/ 才暴露）：
- TestFailedSlotC30Guard 內 class-level PROJECT_ROOT 用 .parent.parent
  原本在根目錄 tests/ 解析正確，搬到 tests/unit/ 後指向 /OpenAver/tests
  改 .parent.parent.parent 統一指向 project root
- find_pattern_in_file helper 末段在 sed 切割時遺失 return violations，補回

無 class 重命名：unit 檔原 71 classes 與根 48 classes 全無重名衝突，可直
append 不需 _Root 後綴。

無業務邏輯改動，純結構搬移 + bug 修正。

Verification:
  - tests/test_frontend_lint.py 不存在 ✓
  - pytest --collect-only tests/unit/test_frontend_lint.py = 450 tests ✓
  - 全套 pytest 2406 passed, 0 failed
  - npm run lint exit 0
  - 無外部硬編碼引用需更新（grep .github/.vscode/*.yml/*.md 確認）
